MemoryProcessor added, CoherencyProcessor fixed
Scaling for L1Invalidations and L2Invalidations fixed. Support for
reading bytes read and written to MC added. Values of L1Invalidations
and L2Invalidations still seem odd.
Hack: currently the MC values and QPI values are read into HwCounters of
the last core. In MemoryProcessor, they are added to the output only if
current core is the last core.